[ARC-589] Separate test extraction from docker test
branch: feature/ARC-589_sbt_test_step
