+++ src/mpyl/steps/test/dockertest.py
@@ -1,11 +1,15 @@
 """ Step that tests the docker image from the target `tester` in Dockerfile-mpl. """
+import shutil
 from logging import Logger
+from pathlib import Path
+
+from python_on_whales import docker
 
 from .. import Step, Meta
-from ..models import Input, Output, ArtifactType
-from ...project import Stage
+from ..models import Input, Output, ArtifactType, input_to_artifact, Artifact
+from ...project import Stage, Project
 from ...utilities.docker import DockerConfig, build, docker_image_tag, docker_file_path
-from ...utilities.junit import to_test_suites, sum_suites, extract_test_results
+from ...utilities.junit import to_test_suites, sum_suites, TEST_OUTPUT_PATH_KEY
 
 
 class TestDocker(Step):
@@ -31,7 +35,7 @@ class TestDocker(Step):
                         file_path=dockerfile, image_tag=tag, target=test_target)
 
         if success:
-            artifact = extract_test_results(project, tag, step_input)
+            artifact = self.extract_test_results(project, tag, step_input)
 
             suite = to_test_suites(artifact)
             summary = sum_suites(suite)
@@ -42,3 +46,15 @@ class TestDocker(Step):
         return Output(success=False,
                       message=f"Tests failed to run for {project.name}. No test results have been recorded.",
                       produced_artifact=None)
+
+    @staticmethod
+    def extract_test_results(project: Project, tag, step_input: Input) -> Artifact:
+        test_result_path = Path(project.target_path, "test_results")
+        shutil.rmtree(test_result_path, ignore_errors=True)
+        Path(test_result_path).mkdir(parents=True, exist_ok=True)
+
+        container_id = docker.create(tag).id
+        docker.copy(f'{container_id}:/{project.test_report_path}/.', test_result_path)
+
+        return input_to_artifact(artifact_type=ArtifactType.JUNIT_TESTS, step_input=step_input,
+                                 spec={TEST_OUTPUT_PATH_KEY: f'{test_result_path}'})
